Added project name to the PROJECT config file.
projectName field is needed for operator-sdk to deduct the metadata name for the
project. When the bundle command reads input files from the stdin and no --project
argument is provided, and projectName is absent the generated metadata name has
invalid pattern: '.v0.18.0'.
This commit adds the projectName field so metadata.name is correctly generated.

Fixes #37