Remove tigrannajaryan as a maintainer
I am no longer able to allocate time to be a maintainer of the operator
due to focusing on other parts of OpenTelemetry.

I am moving myself to the Emeritus section according to the rules introduced in
https://github.com/open-telemetry/community/pull/961

I have also listed approvers and maintainers directly in the README, like we
do in the Collector repos. I will also remove Collector github teams' permissions
from this repo since we have github teams for operator to avoid confusion as to
who has what permissions.