Fix "sidecar.opentelemetry.io/injectec" label value sometimes being longer than 63 characters
Use two new labels:

- `sidecar.opentelemetry.io/injected-namespace`: namespace of the `OpenTelemetryCollector`
- `sidecar.opentelemetry.io/injected-name`: name of the `OpenTelemetryCollector`

Since Kubernetes object names and label values have the same restrictions, we can safely use these two new labels with guarantee that it will not produce errors.

For backwards compatibility, we continue to set the `sidecar.opentelemetry.io/injected` label _only if_ it does not exceed 63 characters. This ensures that it still exists when valid, but does not cause errors when invalid.