fix reconciliation error when prometheusremotewrite exporter is enabled
https://github.com/open-telemetry/opentelemetry-operator/pull/1953
introduced the following error when prometheusremotewrite exporter is
enabled

	failed to reconcile the expected deployments: failed to apply changes: Deployment.apps "cluster-collector" is invalid: spec.template.spec.containers[0].ports[3].containerPort: Required value"

required configuration is:

	exporters:
	  prometheusremotewrite:
	    endpoint: "https://example.com/api/v1/write"

getPrometheusExporterPorts() tries to parse this endpoint into a
host:port pair, because the exporter name contains the string
"prometheus", but prometheusremotewriteexporter's expected `endpoint`
configuration is actually a URL, so that fails. additionally, due to a
missing `continue` if the port parsing fails, an erroneous ContainerPort
entry with port set to 0 is added to the configuration, causing the
error.

Signed-off-by: Lauri Tirkkonen <lauri.tirkkonen@paypay-corp.co.jp>