Add v1alpha2 embedded Target Allocator struct (#2623)
Add and use the embedded Target Allocator struct for v1alpha2. The
changes mostly involve label selectors for Prometheus CR.

Worth noting that this involves a change in semantics for said
selectors, as an empty selector now selects nothing, which is in line
with the standard, but different from v1alpha1.