Added protocol handling for ServicePort and fixed type/value naming inside generic receiver (#2619)
* Fix handling of type

The name is not correctly split before it is checked. Split it first.

Signed-off-by: Oldřich Jedlička <oldium.pro@gmail.com>

* Split parsing of Syslog, TCP Log and UDP Log parsers

The generic implementation does not allow to specify different endpoints
and endpoint-specific protocols, so split the implementation instead of
trying to make the generic implementation a complex catch-all-party.

Signed-off-by: Oldřich Jedlička <oldium.pro@gmail.com>

* Handle port numbers together with protocols

It is legal to have the same protocol number and different protocol, so
handle it accordingly. Also interpret empty protocol as TCP, like the K8s
does.

Signed-off-by: Oldřich Jedlička <oldium.pro@gmail.com>

* Changelog updates

Signed-off-by: Oldřich Jedlička <oldium.pro@gmail.com>

---------

Signed-off-by: Oldřich Jedlička <oldium.pro@gmail.com>