fix: fix container names for sdk, apache, nginx
Found through code inspection, the ApacheHttpd, Nginx and SDK
injectors do not honor their container-names annotations.

This is a breaking change if anyone is using the enablement flag with
container names for these 3 injectors, which I believe would work as
truthy would treat a list of container names as true.

Signed-off-by: Alan Clucas <alan@clucas.org>