[chore] Fix featuregate usage in controller tests (#3490)
We haven't been unsetting feature gates in controller tests after ending
the test, leading to them being enabled for the duration of the test
suite. In one case, a test actually depended on this fact, and I needed
to set the gate in it explicitly.

Also switched to use the gates explicitly vs parsing flags.