feat(ta): allowNamespaces and denyNamespaces
When the target allocator is configured to watch Prometheus custom
resources in the cluster to discover targets, it is currently hard-coded
to require a ClusterRole with a policy rule of listing namespaces. This
prevents usage in environments where configuring ClusterRoles is not
permitted i.e. in namespace-as-a-service setups where only Roles can be
created.

This change introduces two fields in the prometheusCR specification to
allow configuring the namespaces that can be interacted with by the
target allocator.

- allowNamespaces is a list of namespaces for the target allocator to
  watch. If set to an empty list, it will list all namespaces in the
  cluster. This is mutually exclusive with denyNamespaces.  Defaults to
  an empty list.
- denyNamespaces is a list of namespaces for the target allocator to not
  watch. If set to an empty list, it will not exclude any namespaces.
  This is mutually exclusive with allowNamespaces.  Defaults to an empty
  list.

Validation is performed at the admission webhook and TA startup to
validate that both fields are not set at the same time.

A new end to end test was created,
e2e-targetallocator/targetallocator-namespace, to test setting
allowNamespaces and denyNamespaces.

The first part of the test sets up a a collector and TA in a namespace suffixed with
'-top-secret'. The TA in this namespace is configured with the
allowNamespaces to the '-top-secret' namespace. Then a check is
performed to make sure that it can discover the service monitor that is
created by the otel-operator. This setup also doesn't make use of any
ClusterRole as it only creates a Role to perform its task of discovering
targets in its own namespace.

The second part of the test sets up another collector and TA in a
separate namespace without the '-top-secret' namespace with a
ClusterRole. The denyNamespaces is set to the namespace suffixed with
the '-top-secret' namespace, and later we check that we can't discover
the ServiceMonitor that is in the 'top-secret' namespace. We also check
that it can still discover the ServiceMonitor that is in its own
namespace.

Documentation was added to the otel-allocator README on how to set
allowNamespaces to configure the scope that the TA is allowed to
interact with.

Fixes: #3086

Signed-off-by: Charlie Le <charlie_le@apple.com>