Allow for scoping TA to watch namespace
It's currently not possible to deploy the TA without cluster-wide
permisisons.

This change introduces a new env variable to the TA, WATCH_NAMESPACE,
which allows for specifying which namespaces to watch. This approach is
similar to how the opentelemetry-operator can be scoped to watch a
single namespace.

This does mean that cluster-wide resource like node metrics (cAdvisor)
are no longer accessible, but this is acceptable since we only want the
TA to know about targets that exist a specific namespaces.

Fixes: https://github.com/open-telemetry/opentelemetry-operator/issues/3086

Signed-off-by: Charlie Le <charlie_le@apple.com>