Re-try failed upgrades (#3893)
* Re-try failed upgrades

Currently, the operator upgrades all OpenTelemetryCollector CRs on
startup of the operator. If any upgrade fails (e.g. intermittent errors
of the Kubernetes API server), it won't be re-tried until the operator
is restarted.

This commit moves the upgrade step to the reconcile loop. The operator
reconciles all managed instances on startup, and in case of an error,
re-tries the upgrade with exponential backoff.

Additionally, I changed the event type from Error to Warning, because
"Error" is not a valid event type.

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

* add description to changelog

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

* extract upgrade to separate methods, return and re-queue reconciliation on upgrade

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

* move NeedsUpgrade to upgrade package

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

* re-add ManagementState and UpgradeStrategy check to Upgrade function

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

* check NeedsUpgrade() in Upgrade() fn

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

* add unit tests, remove upgrade from status handling

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

* fix tests

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

---------

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>