Return empty config struct in case v1alpha1 config is invalid.
This change is meant to make the conversion webhook to be more resilient
in the face of invalid input.

It is rare, though possible, for an opentelemetrycollector custom
resource to be stored as a v1alpha1 version and with an invalid
configuration. With an older version of the operator running this would
simply result in no collector starting due to an invalid
configuration.

However, once the operator is updated to a new enough version where
v1beta1 is the stored version, this invalid configuration will result in
the conversion webhook failing. After several minutes the operator will
crash.

On a production cluster we had this situation result in the resource
quota controller failing to initialize after a control-plane upgrade.
This resulted in an incident where new jobs could not be created.